// types/visa.ts
export type VisaType =
  | 'NONE'
  | 'A1'
  | 'A2'
  | 'A3'
  | 'B1'
  | 'B2'
  | 'C1'
  | 'C3'
  | 'C4'
  | 'D1'
  | 'D2'
  | 'D3'
  | 'D4'
  | 'D5'
  | 'D6'
  | 'D7'
  | 'D8'
  | 'D9'
  | 'D10'
  | 'E1'
  | 'E2'
  | 'E3'
  | 'E4'
  | 'E5'
  | 'E6'
  | 'E7'
  | 'E74'
  | 'E74R'
  | 'E9'
  | 'E10'
  | 'F1'
  | 'F2'
  | 'F27'
  | 'F2R'
  | 'F3'
  | 'F4'
  | 'F4R'
  | 'F5'
  | 'F6'
  | 'G1'
  | 'H1'
  | 'H2'